karin fix transaction data & currency

+++ MobileBanking/Model/TransactionModel.swift
@@ -15,7 +15,7 @@ struct TransactionResponse: Codable{
 
 struct TransactionData: Codable{
     var transactionId: String
-    var amount: Float
+    var amount: Double
     var transactionDate: String
     var description: String?
     var transactionType: String
